Apartment Station Test and Prefab Creation
Got the current iteration of the station switch to work in the apartment (i.e. spacebar switches every station between past and present at any time). Also created a basic station prefab for ease of use. Also also made a variant of the player prefab that probably won't stick around for long. Was just using it to check things.